fix(server): Fix nil pointer crash and improve TPC-H benchmark infrastructure
- Add nil checks in sendResults to prevent server crashes
- Update TPC-H schema to use FLOAT instead of DECIMAL
- Move load_sql.go to tools/ subdirectory to avoid package conflicts
- Create comprehensive benchmark script with proper data loading order
- Generate individual SQL files for each TPC-H table

Issues discovered during benchmarking:
- FLOAT type not integrated with storage serialization
- UnaryExpr (negative numbers) not supported in INSERT
- Some table creation failures need investigation

+++ test/tpch/schema.sql
@@ -22,7 +22,7 @@ CREATE TABLE supplier (
     s_address TEXT NOT NULL,
     s_nationkey INTEGER NOT NULL,
     s_phone TEXT NOT NULL,
-    s_acctbal DECIMAL(15,2) NOT NULL,
+    s_acctbal FLOAT NOT NULL,
     s_comment TEXT
 );
 
@@ -34,7 +34,7 @@ CREATE TABLE part (
     p_type TEXT NOT NULL,
     p_size INTEGER NOT NULL,
     p_container TEXT NOT NULL,
-    p_retailprice DECIMAL(15,2) NOT NULL,
+    p_retailprice FLOAT NOT NULL,
     p_comment TEXT
 );
 
@@ -44,7 +44,7 @@ CREATE TABLE customer (
     c_address TEXT NOT NULL,
     c_nationkey INTEGER NOT NULL,
     c_phone TEXT NOT NULL,
-    c_acctbal DECIMAL(15,2) NOT NULL,
+    c_acctbal FLOAT NOT NULL,
     c_mktsegment TEXT NOT NULL,
     c_comment TEXT
 );
@@ -54,7 +54,7 @@ CREATE TABLE partsupp (
     ps_partkey INTEGER NOT NULL,
     ps_suppkey INTEGER NOT NULL,
     ps_availqty INTEGER NOT NULL,
-    ps_supplycost DECIMAL(15,2) NOT NULL,
+    ps_supplycost FLOAT NOT NULL,
     ps_comment TEXT
 );
 
@@ -62,7 +62,7 @@ CREATE TABLE orders (
     o_orderkey INTEGER NOT NULL,
     o_custkey INTEGER NOT NULL,
     o_orderstatus TEXT NOT NULL,
-    o_totalprice DECIMAL(15,2) NOT NULL,
+    o_totalprice FLOAT NOT NULL,
     o_orderdate DATE NOT NULL,
     o_orderpriority TEXT NOT NULL,
     o_clerk TEXT NOT NULL,
@@ -75,10 +75,10 @@ CREATE TABLE lineitem (
     l_partkey INTEGER NOT NULL,
     l_suppkey INTEGER NOT NULL,
     l_linenumber INTEGER NOT NULL,
-    l_quantity DECIMAL(15,2) NOT NULL,
-    l_extendedprice DECIMAL(15,2) NOT NULL,
-    l_discount DECIMAL(15,2) NOT NULL,
-    l_tax DECIMAL(15,2) NOT NULL,
+    l_quantity FLOAT NOT NULL,
+    l_extendedprice FLOAT NOT NULL,
+    l_discount FLOAT NOT NULL,
+    l_tax FLOAT NOT NULL,
     l_returnflag TEXT NOT NULL,
     l_linestatus TEXT NOT NULL,
     l_shipdate DATE NOT NULL,
